Use EventIdPropertyCache in SerilogLogger; Create EventIdPropertyCache test class.
