+++ src/Serilog.Extensions.Logging/Extensions/Logging/SerilogLogger.cs
@@ -29,12 +29,12 @@ class SerilogLogger : FrameworkLogger
     readonly SerilogLoggerProvider _provider;
     readonly ILogger _logger;
 
-    static readonly CachingMessageTemplateParser MessageTemplateParser = new();
+    /// Per-ILogger LogEventProperty cache for EventId.
+    /// Each ILogger has its own cache.
+    /// Each { ILogger, EventId.Id } pair is assumed to be unique.
+    readonly EventIdPropertyCache _eventPropertyCache = new (48);
 
-    // It's rare to see large event ids, as they are category-specific
-    static readonly LogEventProperty[] LowEventIdValues = Enumerable.Range(0, 48)
-        .Select(n => new LogEventProperty("Id", new ScalarValue(n)))
-        .ToArray();
+    static readonly CachingMessageTemplateParser MessageTemplateParser = new();
 
     public SerilogLogger(
         SerilogLoggerProvider provider,
@@ -155,7 +155,7 @@ class SerilogLogger : FrameworkLogger
         }
 
         if (eventId.Id != 0 || eventId.Name != null)
-            properties.Add(CreateEventIdProperty(eventId));
+            properties.Add(this._eventPropertyCache.GetOrCreateProperty(in eventId));
 
         var (traceId, spanId) = Activity.Current is { } activity ?
             (activity.TraceId, activity.SpanId) :
@@ -171,26 +171,5 @@ class SerilogLogger : FrameworkLogger
         if (formatter != null)
             stateObj = formatter(state, null);
         return stateObj ?? state;
-    }
-
-    internal static LogEventProperty CreateEventIdProperty(EventId eventId)
-    {
-        var properties = new List<LogEventProperty>(2);
-
-        if (eventId.Id != 0)
-        {
-            if (eventId.Id >= 0 && eventId.Id < LowEventIdValues.Length)
-                // Avoid some allocations
-                properties.Add(LowEventIdValues[eventId.Id]);
-            else
-                properties.Add(new LogEventProperty("Id", new ScalarValue(eventId.Id)));
-        }
-
-        if (eventId.Name != null)
-        {
-            properties.Add(new LogEventProperty("Name", new ScalarValue(eventId.Name)));
-        }
-
-        return new LogEventProperty("EventId", new StructureValue(properties));
     }
 }
